Convert logging to Typescript
